React DnD works withour Mutation errors; still need to fix rerender after update
